pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "SlateGallery"
version = "1.0.0"
description = "PySide6-based photo gallery generator with EXIF metadata extraction"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PySide6>=6.4",
    "Pillow>=10.0.0",
    "Jinja2>=3.1.0",
    "piexif>=1.1.3",
    "typing-extensions>=4.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "basedpyright>=1.20.0",
    "pytest>=8.0.0",
    "pytest-qt>=4.4.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [".git", "__pycache__", "venv", ".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # bugbear (likely bugs)
    "SIM",    # simplify
    "RUF",    # ruff-specific
    "PIE",    # misc linting
    "C4",     # comprehension improvements
]
ignore = ["E501"]

[tool.basedpyright]
pythonVersion = "3.9"
typeCheckingMode = "strict"
include = ["src", "run_slate_gallery.py"]
exclude = [".git", "venv", ".venv", "archive", "__pycache__"]
venvPath = "."
venv = ".venv"
extraPaths = ["src"]
reportMissingImports = true
reportMissingTypeStubs = false
reportExplicitAny = "warning"
reportAny = "warning"
reportAttributeAccessIssue = "warning"
reportUninitializedInstanceVariable = "warning"
reportOptionalMemberAccess = "warning"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v"

[tool.coverage.run]
source = ["src"]
omit = ["*/__pycache__/*", "*/tests/*"]

[tool.coverage.report]
show_missing = true
